Improve codepoint name logging in the ospf2 and ospf3 serializer
