+++ code/lspgen/src/lspgen_lsdb.c
@@ -752,28 +752,6 @@ lsdb_format_isis_attr(struct lsdb_attr_ *attr)
     return buf;
 }
 
-/*
- * Link State Attributes / name mappings
- */
-struct keyval_ ospf_attr_names[] = {
-    { OSPF_MSG_LSUPDATE,		"LS-Update" },
-    { OSPF_LSA_ROUTER,			"Router-LSA" },
-    { OSPF_LSA_EXTERNAL,		"External-LSA" },
-    { OSPF_LSA_EXTERNAL6,		"External-LSA" },
-    { OSPF_LSA_INTRA_AREA_PREFIX,       "Intra-Area-Prefix-LSA" },
-    { OSPF_LSA_OPAQUE_AREA_RI,		"Opaque-LSA-RI" },
-    { OSPF_LSA_OPAQUE_AREA_EP,		"Opaque-LSA-EP" },
-    { OSPF_ROUTER_LSA_LINK_PTP,		"ptp-link" },
-    { OSPF_ROUTER_LSA_LINK_STUB,	"ptp-stub" },
-    { OSPF_TLV_HOSTNAME,		"Hostname" },
-    { OSPF_TLV_SID_LABEL_RANGE,		"SID/Label-Range" },
-    { OSPF_TLV_EXTENDED_PREFIX,		"Extended-Prefix" },
-    { OSPF_TLV_EXTENDED_PREFIX_RANGE,	"Extended-Prefix-Range" },
-    { OSPF_SUBTLV_PREFIX_SID,		"Prefix-SID" },
-    { OSPF_IA_PREFIX_LSA_PREFIX,	"Prefix" },
-    { 0, NULL}
-};
-
 char *
 lsdb_format_ospf_attr(struct lsdb_attr_ *attr)
 {
